+++ lib/bind-keys.js
