+++ lib/board.js
@@ -1,7 +1,7 @@
 const _             = require("lodash");
-const boardSettings = require("./game-settings/board_settings");
-const LoadingArea   = require("./loading_area");
-const ConveyorArea  = require("./conveyor_area");
+const boardSettings = require("./game-settings/board-settings");
+const LoadingArea   = require("./loading-area");
+const ConveyorArea  = require("./conveyor-area");
 const PlayerEvents  = require("./player-events");
 const extend        = require("./util/extend");
 const incGet        = require("./util/inc-get");
