+++ lib/cell.js
@@ -1,4 +1,4 @@
-const cellSettings = require("./game-settings/cell_settings");
+const cellSettings = require("./game-settings/cell-settings");
 const extend       = require("./util/extend");
 const incGet       = require("./util/inc-get");
