+++ lib/color-blob.js
@@ -1,4 +1,4 @@
-const colorBlobSettings = require("./game-settings/color_blob_settings");
+const colorBlobSettings = require("./game-settings/color-blob-settings");
 
 function ColorBlob(laneIndex, settings) {
   this.settings = settings || colorBlobSettings();
