+++ lib/conveyor-area.js
@@ -1,6 +1,6 @@
 const _ = require("lodash");
 
-const conveyorAreaSettings = require("./game-settings/conveyor_area_settings");
+const conveyorAreaSettings = require("./game-settings/conveyor-area-settings");
 const Lane = require("./lane");
 
 function ConveyorArea(settings) {
