+++ lib/conveyor-belt.js
@@ -1,5 +1,5 @@
 const _                    = require("lodash");
-const conveyorBeltSettings = require("./game-settings/conveyor_belt_settings");
+const conveyorBeltSettings = require("./game-settings/conveyor-belt-settings");
 const Cell                 = require("./cell");
 const extend               = require("./util/extend");
 const incGet               = require("./util/inc-get");
