+++ lib/game-settings/board-settings.js
@@ -1,6 +1,6 @@
-const getLaneStarts          = require("../util/get_lane_starts");
-const getvConveyorBeltStarts = require("../util/get_conveyor_belt_starts");
-const getvColorBlobStarts    = require("../util/get_color_blob_starts");
+const getLaneStarts          = require("../util/get-lane-starts");
+const getvConveyorBeltStarts = require("../util/get-conveyor-belt-starts");
+const getvColorBlobStarts    = require("../util/get-color-blob-starts");
 const getLaneWorkerStarts    = require("../util/get-lane-worker-starts");
 
 function boardSettings() {
