+++ lib/game-settings/cell-settings.js
@@ -1,4 +1,4 @@
-const getBoardSettings = require("./board_settings");
+const getBoardSettings = require("./board-settings");
 
 function cellSettings() {
   var boardSettings = getBoardSettings();
