+++ lib/game-settings/color-blob-settings.js
@@ -1,4 +1,4 @@
-const getBoardSettings = require("./board_settings");
+const getBoardSettings = require("./board-settings");
 
 function colorBlobSettings() {
   var boardSettings = getBoardSettings();
