+++ lib/game-settings/conveyor-area-settings.js
@@ -1,4 +1,4 @@
-const getBoardSettings = require("./board_settings");
+const getBoardSettings = require("./board-settings");
 
 function conveyorAreaSettings() {
   var boardSettings = getBoardSettings();
