+++ lib/game-settings/conveyor-belt-settings.js
@@ -1,4 +1,4 @@
-const getBoardSettings = require("./board_settings");
+const getBoardSettings = require("./board-settings");
 
 function conveyorBeltSettings() {
   var boardSettings = getBoardSettings();
