+++ lib/game-settings/lane-worker-settings.js
@@ -1,4 +1,4 @@
-const getBoardSettings = require("./board_settings");
+const getBoardSettings = require("./board-settings");
 
 function laneWorkerSettings() {
   var boardSettings = getBoardSettings();
