+++ lib/game-settings/loading-area-settings.js
@@ -1,4 +1,4 @@
-const getBoardSettings = require("./board_settings");
+const getBoardSettings = require("./board-settings");
 
 function loadingAreaSettings() {
   var boardSettings = getBoardSettings();
