+++ lib/game-settings/selected-lane-settings.js
@@ -1,4 +1,4 @@
-const getBoardSettings = require("./board_settings");
+const getBoardSettings = require("./board-settings");
 
 function selectedLaneSettings() {
   var boardSettings = getBoardSettings();
