+++ lib/index.js
@@ -1,6 +1,6 @@
 const $ = require("jquery");
 const Board = require("./board");
-const bindKeys = require("./bind_keys");
+const bindKeys = require("./bind-keys");
 const Setup = require("./setup");
