+++ lib/lane-worker.js
@@ -1,4 +1,4 @@
-const laneWorkerSettings = require("./game-settings/lane_worker_settings");
+const laneWorkerSettings = require("./game-settings/lane-worker-settings");
 const extend       = require("./util/extend");
 const incGet       = require("./util/inc-get");
