+++ lib/lane.js
@@ -1,8 +1,8 @@
 const _            = require("lodash");
-const laneSettings = require("./game-settings/lane_settings");
+const laneSettings = require("./game-settings/lane-settings");
 
-const ConveyorBelt = require("./conveyor_belt");
-const LaneWorker   = require("./lane_worker");
+const ConveyorBelt = require("./conveyor-belt");
+const LaneWorker   = require("./lane-worker");
 
 function Lane(laneIndex, settings) {
   this.settings = settings || laneSettings();
