+++ lib/loading-area.js
@@ -1,4 +1,4 @@
-const loadingAreaSettings = require("./game-settings/loading_area_settings");
+const loadingAreaSettings = require("./game-settings/loading-area-settings");
 const SelectedLane = require("./selected-lane");
 
 function LoadingArea(settings) {
