+++ lib/util/get-color-blob-starts.js
