+++ test/blob-engine-test.js
@@ -1,0 +1,14 @@
+const _      = require("lodash");
+const chai   = require('chai');
+const expect = chai.expect;
+const Board  = require('../lib/board');
+const BlobEngine  = require('../lib/blob-engine');
+
+describe('blob engine object', function () {
+  let game =  new Board();
+  let blobEngine = new BlobEngine();
+
+  context("")
+
+
+});
